Refactor base config settings.
